requirements.txt
# Dependências do Projeto Princesa Ana Paula
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
PyMySQL==1.1.0